% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_type}
\alias{check_type}
\title{Verify that a value input is of appropriate type}
\usage{
check_type(x, type)
}
\arguments{
\item{x}{Focal input.}

\item{type}{Type of input that \code{x} should be. Presently available 
are \code{"character"} (standard) and \code{"integer"} (conformable to
an integer if not explicitly classed as such)/}
}
\value{
\code{NULL} if \code{x} is proper, throwing an error otherwise.
}
\description{
Throws an error if \code{x} isn't proper, based on the 
 \code{type}.
}
\examples{
 names <- "sam"
 check_type(x = names, type = "character")
 names <- 2019
 #check_type(x = names, type = "character") # throws error
 years <- 2019
 check_type(x = years, type = "integer")
 years <- "sam"
 #check_type(x = years, type = "integer") # throws error

}
