% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{language_codes}
\alias{language_codes}
\title{Print Google Language Codes}
\format{An object of class \code{character} of length 81.}
\usage{
language_codes
}
\description{
Print Google Language Codes
}
\seealso{
\link[translateR]{getGoogleLanguages}
}
\keyword{datasets}
