% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_locale}
\alias{get_locale}
\title{Get the language and location of a system locale}
\usage{
get_locale(category = "LC_TIME", locale_asis = FALSE)
}
\arguments{
\item{category}{\code{character} string of the locale environmental variable
to use, passed directly to \code{\link[base]{Sys.getlocale}}.}

\item{locale_asis}{\code{boolean} for returning \emph{language} and
\emph{location} from \link[base]{Sys.getlocale} as it is. Defult to
FALSE.}
}
\value{
\code{character} vector with elements \code{"language"} and
  \code{"location"}.
}
\description{
OS-flexible approach to determining the system locale with
  respect to language and location.
}
\examples{
 get_locale()

}
