% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_locale}
\alias{get_locale}
\title{Get the language and location of a system locale}
\usage{
get_locale(category = "LC_TIME")
}
\arguments{
\item{category}{\code{character} string of the locale environmental 
variable to use, passed directly to \code{\link[base]{Sys.getlocale}}.}
}
\value{
\code{character} vector with elements \code{"language"} and 
 \code{"location"}.
}
\description{
OS-flexible approach to determining the system locale with 
 respect to language and location.
}
\examples{
 get_locale()

}
